Move tabs in shared components folder.
