+++ coral/src/app/features/approvals/components/ApprovalResourceTabs.tsx
